Rename mesher file
