+++ src/game/world/GreedyMesher2d.test.ts
@@ -1,5 +1,5 @@
 import { describe, test, expect } from 'vitest'
-import { GreedyMesher2d } from './GreedyMesher2d_'
+import { GreedyMesher2d } from './GreedyMesher2d'
 
 const dataGetterFn = (data: number[][]) => (x: number, y: number) => data[data.length - y - 1][x]
