+++ src/game/world/GreedyMesher2d.ts
